utils/utilities/pyproject.toml
[project]
name = "utilities"
version = "0.1.0"
description = "Shared utilities"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
    "sqlalchemy-data-model-visualizer>=0.1.3",
    "sqlalchemy>=2.0.46",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
